Split executables.py into one module per command
